Correciones servicios y manejo de excepciones

+++ cmd/api/domain/service/movie_delete_service.go
@@ -3,6 +3,7 @@ package service
 import (
 	"errors"
 
+	"github.com/ederj98/movies-microservice/cmd/api/domain/exception"
 	"github.com/ederj98/movies-microservice/cmd/api/domain/port"
 	"github.com/ederj98/movies-microservice/pkg/logger"
 )
@@ -16,6 +17,13 @@ type MovieDeleteService struct {
 }
 
 func (movieDeleteService *MovieDeleteService) Delete(id int64) (err error) {
+
+	_, exist := movieDeleteService.MovieRepository.Find(id)
+	if exist != nil {
+		err = exception.DataNotFound{ErrMessage: errorNotFoundRepository}
+		logger.Error(errorRepository, err)
+		return err
+	}
 
 	err = movieDeleteService.MovieRepository.Delete(id)
